agent/requirements.txt
# Core
bedrock-agentcore>=0.1.0
boto3>=1.35.0

# Framework
strands-agents>=0.1.0

# Server
uvicorn>=0.30.0
fastapi>=0.110.0

# Observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
